==> 4_TFIDF.ipynb <==
## Resources
- https://github.com/boudinfl/pke 
- https://www.analyticsvidhya.com/blog/2021/12/how-to-extract-key-phrases-using-tfidf-with-python/
- https://towardsdatascience.com/mining-an-economic-news-article-using-pre-trained-language-models-f75af041ecf0

import time
import pandas as pd
import numpy as np
# import nltk
# import gensim
# from gensim import corpora
import matplotlib.pyplot as plt
%matplotlib inline

import nltk
nltk.download('punkt')

from google.colab import drive
drive.mount('/drive')

# TFIDF - Test Data

lda_df.info()

test_df = pd.read_csv('/drive/My Drive/NLP_Project/data/no_chicago_less300.csv')
print(test_df.shape)
test_df.head(3)

(test_df['clean_text'].isna()).index

test_df.dropna(subset=['clean_text'], axis=0, inplace=True)

test_df = test_df[test_df.leng > 200]

from sklearn.feature_extraction.text import TfidfVectorizer

#unigrams

corpus = test_df['clean_text']
tfidf=TfidfVectorizer().fit(corpus)
count_tokens=tfidf.get_feature_names_out()
article_vect = tfidf.transform(corpus)
#The following line is the solution:
# out = {article_vect.tocsr().sum(axis=0)}
tf_idf_DF=pd.DataFrame(data=article_vect.tocsr().sum(axis=0),columns=count_tokens)
tf_idf_DF = tf_idf_DF.T.sort_values(ascending=False, by=[0])

# tf_idf_DF['word'] = tf_idf_DF.index
tf_idf_DF['tf-idf'] = tf_idf_DF[0]
tf_idf_DF = tf_idf_DF.drop([0],axis=1)


def top_tfidf_terms(text):
  # print(text)
  top_terms = {}
  tokens = text.split()
  for word in tf_idf_DF.index:
    if word in tokens:
      # tup = (word,tf_idf_DF.loc[word][0])
      # top_terms.append(tup)
      top_terms[word] = tf_idf_DF.loc[word][0]

    if len(top_terms) == 5:
      break
  
  return [top_terms]

%%time
test_df['tfidf_uni'] = np.NaN
for idx,row in test_df.iterrows():
  # print(idx)
  test_df.loc[idx,'tfidf_uni'] = top_tfidf_terms(row.loc['clean_text'])

from sklearn.feature_extraction.text import TfidfVectorizer

#bigrams

corpus = test_df['clean_text']
tfidf=TfidfVectorizer(ngram_range=(2,2)).fit(corpus)
n_tokens=tfidf.get_feature_names_out()
article_vect = tfidf.transform(corpus)
#The following line is the solution:
# out = {article_vect.tocsr().sum(axis=0)}
tf_idf_DF=pd.DataFrame(data=article_vect.tocsr().sum(axis=0),columns=n_tokens)
tf_idf_DF = tf_idf_DF.T.sort_values(ascending=False, by=[0])

# tf_idf_DF['word'] = tf_idf_DF.index
tf_idf_DF['tf-idf'] = tf_idf_DF[0]
tf_idf_DF = tf_idf_DF.drop([0],axis=1)


def top_tfidf_bi_terms(text):
  # print(text)
  top_terms = {}
  tfidf=TfidfVectorizer(ngram_range=(2,2)).fit([text])
  bi_tokens=tfidf.get_feature_names_out()
  # print(bi_tokens)
  # tokens = text.split()
  for word in tf_idf_DF.index:
    if word in bi_tokens:
      # tup = (word,tf_idf_DF.loc[word][0])
      # top_terms.append(tup)
      top_terms[word] = tf_idf_DF.loc[word][0]

    if len(top_terms) == 5:
      # print(top_terms)
      break
  
  return [top_terms]

%%time
test_df['tfidf_bi_withdf'] = np.NaN
for idx,row in test_df.iterrows():
  # print(idx)
  test_df.loc[idx,'tfidf_bi_withdf'] = top_tfidf_bi_terms(row.loc['clean_text'])

from sklearn.feature_extraction.text import TfidfVectorizer

#triigrams

corpus = test_df['clean_text']
tfidf=TfidfVectorizer(ngram_range=(3,3)).fit(corpus)
n_tokens=tfidf.get_feature_names_out()
article_vect = tfidf.transform(corpus)
#The following line is the solution:
# out = {article_vect.tocsr().sum(axis=0)}
tf_idf_DF=pd.DataFrame(data=article_vect.tocsr().sum(axis=0),columns=n_tokens)
tf_idf_DF = tf_idf_DF.T.sort_values(ascending=False, by=[0])

# tf_idf_DF['word'] = tf_idf_DF.index
tf_idf_DF['tf-idf'] = tf_idf_DF[0]
tf_idf_DF = tf_idf_DF.drop([0],axis=1)

def top_tfidf_n_terms(text,n):
  # print(text)
  top_terms = {}
  tfidf=TfidfVectorizer(ngram_range=(n,n)).fit([text])
  ngram_tokens=tfidf.get_feature_names_out()
  # print(bi_tokens)
  # tokens = text.split()
  for word in tf_idf_DF.index:
    if word in ngram_tokens:
      # tup = (word,tf_idf_DF.loc[word][0])
      # top_terms.append(tup)
      top_terms[word] = tf_idf_DF.loc[word][0]

    if len(top_terms) == 5:
      # print(top_terms)
      break
  
  return [top_terms]

%%time
test_df['tfidf_tri_withdf'] = np.NaN
for idx,row in test_df.iterrows():
  # print(idx)
  test_df.loc[idx,'tfidf_tri_withdf'] = top_tfidf_n_terms(row.loc['clean_text'],3)

tf_idf_DF.head(3)

tf_idf_DF.index

n_tokens

test_df.head(3)

test_df.to_csv('/drive/My Drive/NLP_Project/data/tfidf_testdata2.csv')

# TFIDF - Full data

df0 = pd.read_json('https://storage.googleapis.com/msca-bdp-data-open/news/news_final_project.json', orient='records', lines=True)

### Droping unwanted

df0 = df0.drop('language',axis='columns')

df0.shape

df0['length'] = df0['text'].apply(len)
df0 = df0[df0.length > 250]
df0.shape

df0.drop_duplicates(subset=['title'], keep='first', inplace=True)
df0.shape

df0.head(2)

### Preprocessing Script

import nltk.data
from nltk.corpus import stopwords
import re
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
from nltk.stem import WordNetLemmatizer
from nltk import word_tokenize, sent_tokenize, pos_tag

def convert_to_lower(text):
  return text.lower()
#------------------------------------------------------------------------------
def remove_numbers(text):
  text = re.sub(r'\d+' , '', text)
  return text
#------------------------------------------------------------------------------
def remove_http(text):
  text = re.sub("https?:\/\/t.co\/[A-Za-z0-9]*", ' ', text)
  return text
#------------------------------------------------------------------------------
def remove_short_words(text):
  text = re.sub(r'\b\w{1,2}\b', '', text)
  return text
#------------------------------------------------------------------------------
def remove_punctuation(text):
     punctuations = '''!()[]{};«№»:'"\,`<>./?@=#$-(%^)+&[*_]~'''
     no_punct = ""
     
     for char in text:
        if char not in punctuations:
            no_punct = no_punct + char
     return no_punct
#------------------------------------------------------------------------------
def remove_white_space(text):
  text = text.strip()
  return text
#------------------------------------------------------------------------------
def toknizing(text):
  stp = my_stopwords
  #stops = set(stopwords.words('english'))
  stop_words = set(stp)
  tokens = word_tokenize(text)
  ## Remove Stopwords from tokens
  result = [i for i in tokens if not i in stop_words]
  return result

#########################################
def preprocess_text(text):
    # 1. Tokenise to alphabetic tokens
    text = remove_numbers(text)
    text = remove_http(text)
    text = remove_punctuation(text)
    text = convert_to_lower(text)
    text = remove_white_space(text)
    text = remove_short_words(text)
    tokens = toknizing(text)
    # 2. POS tagging
    pos_map = {'J': 'a', 'N': 'n', 'R': 'r', 'V': 'v'}
    pos_tags_list = pos_tag(tokens)
    return tokens

import nltk
nltk.download('stopwords')
nltk.download('wordnet')
my_stopwords= ["...",".","0o", "0s", "3a", "3b", "3d", "6b", "6o", "a", "a1", "a2", "a3", 
               "a4", "ab","aab",'aaab', "able", "about", "above", "abst", "ac", "accordance",
               "according", "accordingly", "across", "act", "actually", "ad", "added", "adj",
               "ae", "af", "affected", "affecting", "affects", "after", "afterwards", "ag",
               "again", "against", "ah", "ain", "ain't", "aj", "al", "all", "allow", "allows",
               "almost", "alone", "along", "already", "also", "although", "always", "am", "among",
               "amongst", "amoungst", "amount", "an", "and", "announce", "another", "any", "anybody",
               "anyhow", "anymore", "anyone", "anything", "anyway", "anyways", "anywhere", "ao", "ap",
               "apart", "apparently", "appear", "appreciate", "appropriate", "approximately", "ar", "are",
               "aren", "arent", "aren't", "arise", "around", "as", "a's", "aside", "ask", "asking", "associated",
               "at", "au", "auth", "av", "available", "aw", "away", "awfully", "ax", "ay", "az", "b", "b1", "b2", "b3", "ba", "back",
               "bc", "bd", "be", "became", "because", "become", "becomes", "becoming", "been", "before", "beforehand", "begin", "beginning",
               "beginnings", "begins", "behind", "being", "believe", "below", "beside", "besides", "best", "better", "between", "beyond",
               "bi", "bill", "biol", "bj", "bk", "bl", "bn", "both", "bottom", "bp", "br", "brief", "briefly", "bs", "bt", "bu", "but",
               "bx", "by", "c", "c1", "c2", "c3", "ca", "call", "came", "can", "cannot", "cant", "can't", "cause", "causes", "cc", "cd",
               "ce", "certain", "certainly", "cf", "cg", "ch", "changes", "ci", "cit", "cj", "cl", "clearly", "cm", "c'mon", "cn", "co",
               "com", "come", "comes", "con", "concerning", "consequently", "consider", "considering", "contain", "containing", "contains",
               "corresponding", "could", "couldn", "couldnt", "couldn't", "course", "cp", "cq", "cr", "cry", "cs", "c's", "ct", "cu",
               "currently", "cv", "cx", "cy", "cz", "d", "d2", "da", "date", "dc", "dd", "de", "definitely", "describe", "described",
               "despite", "detail", "df", "di", "did", "didn", "didn't", "different", "dj", "dk", "dl", "do", "does", "doesn", "doesn't",
               "doing", "don", "done", "don't", "down", "downwards", "dp", "dr", "ds", "dt", "du", "due", "during", "dx", "dy", "e", "e2",
               "e3", "ea", "each", "ec", "ed", "edu", "ee", "ef", "effect", "eg", "ei", "eight", "eighty", "either", "ej", "el", "eleven",
               "else", "elsewhere", "em", "empty", "en", "end", "ending", "enough", "entirely", "eo", "ep", "eq", "er", "es", "especially",
               "est", "et", "et-al", "etc", "eu", "ev", "even", "ever", "every", "everybody", "everyone", "everything", "everywhere", "ex",
               "exactly", "example", "except", "ey", "f", "f2", "fa", "far", "fc", "few", "ff", "fi", "fifteen", "fifth", "fify", "fill", "find",
               "fire", "first", "five", "fix", "fj", "fl", "fn", "fo", "followed", "following", "follows", "for", "former", "formerly",
               "forth", "forty", "found", "four", "fr", "from", "front", "fs", "ft", "fu", "full", "further", "furthermore", "fy", "g",
               "ga", "gave", "ge", "get", "gets", "getting", "gi", "give", "given", "gives", "giving", "gj", "gl", "go", "goes", "going",
               "gone", "got", "gotten", "gr", "greetings", "gs", "gy", "h", "h2", "h3", "had", "hadn", "hadn't", "happens", "hardly", "has",
               "hasn", "hasnt", "hasn't", "have", "haven", "haven't", "having", "he", "hed", "he'd", "he'll", "hello", "help", "hence", "her",
               "here", "hereafter", "hereby", "herein", "heres", "here's", "hereupon", "hers", "herself", "hes", "he's", "hh", "hi", "hid",
               "him", "himself", "his", "hither", "hj", "ho", "home", "hopefully", "how", "howbeit", "however", "how's", "hr", "hs", 
               "hu", "hundred", "hy", "i", "i2", "i3", "i4", "i6", "i7", "i8", "ia", "ib", "ibid", "ic", "id", "i'd", "ie", "if", "ig", "ignored",
               "ih", "ii", "ij", "il", "i'll", "im", "i'm", "immediate", "immediately", "importance", "important", "in", "inasmuch", "inc",
               "indeed", "index", "indicate", "indicated", "indicates", "information", "inner", "insofar", "instead", "interest", "into",
               "invention", "inward", "io", "ip", "iq", "ir", "is", "isn", "isn't", "it", "itd", "it'd", "it'll", "its", "it's", "itself",
               "iv", "i've", "ix", "iy", "iz", "j", "jj", "jr", "js", "jt", "ju", "just", "k", "ke", "keep", "keeps", "kept", "kg", "kj",
               "km", "know", "known", "knows", "ko", "l", "l2", "la", "largely", "last", "lately", "later", "latter", "latterly", "lb", "lc",
               "le", "least", "les", "less", "lest", "let", "lets", "let's", "lf", "like", "liked", "likely", "line", "little", "lj", "ll",
               "ll", "ln", "lo", "look", "looking", "looks", "los", "lr", "ls", "lt", "ltd", "m", "m2", "ma", "made", "mainly", "make", "makes",
               "many", "may", "maybe", "me", "mean", "means", "meantime", "meanwhile", "merely", "mg", "might", "mightn", "mightn't", "mill",
               "million", "mine", "miss", "ml", "mn", "mo", "more", "moreover", "most", "mostly", "move", "mr", "mrs", "ms", "mt", "mu", "much",
               "mug", "must", "mustn", "mustn't", "my", "myself", "n", "n2", "na", "name", "namely", "nay", "nc", "nd", "ne", "near", "nearly",
               "necessarily", "necessary", "need", "needn", "needn't", "needs", "neither", "never", "nevertheless", "new", "next", "ng", "ni",
               "nine", "ninety", "nj", "nl", "nn", "no", "nobody", "non", "none", "nonetheless", "noone", "nor", "normally", "nos", "not",
               "noted", "nothing", "novel", "now", "nowhere", "nr", "ns", "nt", "ny", "o", "oa", "ob", "obtain", "obtained", "obviously",
               "oc", "od", "of", "off", "often", "og", "oh", "oi", "oj", "ok", "okay", "ol", "old", "om", "omitted", "on", "once", "one", "ones",
               "only", "onto", "oo", "op", "oq", "or", "ord", "os", "ot", "other", "others", "otherwise", "ou", "ought", "our", "ours",
               "ourselves", "out", "outside", "over", "overall", "ow", "owing", "own", "ox", "oz", "p", "p1", "p2", "p3", "page", "pagecount",
               "pages", "par", "part", "particular", "particularly", "pas", "past", "pc", "pd", "pe", "per", "perhaps", "pf", "ph", "pi", "pj",
               "pk", "pl", "placed", "please", "plus", "pm", "pn", "po", "poorly", "possible", "possibly", "potentially", "pp", "pq", "pr",
               "predominantly", "present", "presumably", "previously", "primarily", "probably", "promptly", "proud", "provides", "ps", "pt",
               "pu", "put", "py", "q", "qj", "qu", "que", "quickly", "quite", "qv", "r", "r2", "ra", "ran", "rather", "rc", "rd", "re",
               "ref", "refs", "regarding", "regardless", "regards", "related",
               "relatively", "rf", "rh", "ri", "right",
               "rj", "rl", "rm", "rn", "ro", "rq", "rr", "rs", "rt", "ru", "run", "rv", "ry", "s", "s2", "sa", "said", "same", "saw", "say",
               "saying", "says", "sc", "sd", "se", "sec", "second", "secondly", "section", "see", "seeing", "seem", "seemed", "seeming",
               "seems", "seen", "self", "selves", "sent",  "seven", "several", "sf", "shall", "shan",
               "shan't", "she", "shed", "she'd", "she'll", "shes", "she's", "should", "shouldn", "shouldn't", "should've", "show", "showed",
               "shown", "showns", "shows", "si", "side", "significant", "significantly", "similar", "similarly", "since", "sincere",
               "six", "sixty", "sj", "sl", "slightly", "sm", "sn", "so", "some", "somebody", "somehow", "someone", "somethan", "something",
               "sometime", "sometimes", "somewhat", "somewhere", "soon", "sorry", "sp", "specifically", "specified", "specify", "specifying",
               "sq", "sr", "ss", "st", "still", "stop", "strongly", "sub", "such", 
               "suggest", "sup", "sure", "sy", "system", "sz", "t", "t1", "t2", "t3", "take", "taken", "taking", "tb", "tc", "td", "te",
               "tell", "ten", "tends", "tf", "th", "than", "thank", "thanks", "thanx", "that", "that'll", "thats", "that's", "that've",
               "the", "their", "theirs", "them", "themselves", "then", "thence", "there", "thereafter", "thereby", "thered",
               "therefore", "therein", "there'll", "thereof", "therere", "theres", "there's", "thereto", "thereupon", "there've", "these",
               "they", "theyd", "they'd", "they'll", "theyre", "they're", "they've", "thickv", "thin", "think", "third", "this", "thorough",
               "thoroughly", "those", "thou", "though", "thoughh", "thousand", "three", "throug", "through", "throughout", "thru",
               "thus", "ti", "til", "tip", "tj", "tl", "tm", "tn", "to", "together", "too", "took", "top", "toward", "towards", "tp",
               "tq", "tr", "tried", "tries", "truly", "try", "trying", "ts", "t's", "tt", "tv", "twelve", "twenty", "twice", "two", "tx",
               "u", "u201d", "ue", "ui", "uj", "uk", "um", "un", "under", "unfortunately", "unless", "unlike", "unlikely", "until", "unto",
               "uo", "up", "upon", "ups", "ur", "us", "use", "used", "useful", "usefully", "usefulness", "uses", "using", "usually",
               "ut", "v", "va", "value", "various", "vd", "ve", "ve", "very", "via", "viz", "vj", "vo", "vol", "vols", "volumtype", "vq",
               "vs", "vt", "vu", "w", "wa", "want", "wants", "was", "wasn", "wasnt", "wasn't", "way", "we", "wed", "we'd", "welcome",
               "well", "we'll", "well-b", "went", "were", "we're", "weren", "werent", "weren't", "we've", "what", "whatever", "what'll",
               "whats", "what's", "when", "whence", "whenever", "when's", "where", "whereafter", "whereas", "whereby", "wherein", "wheres",
               "where's", "whereupon", "wherever", "whether", "which", "while", "whim", "whither", "who", "whod", "whoever", "whole",
               "who'll", "whom", "whomever", "whos", "who's", "whose", "why", "why's", "wi", "widely", "will", "willing", "wish", "with",
               "within", "without", "wo", "won", "wont", "won't", "words", "would", "wouldn", "wouldnt", "wouldn't",
               "www", "x", "x1", "x2", "x3", "xf", "xi", "xj", "xk", "xl", "xn", "xo", "xs", "xt", "xv", "xx", "y", "y2", "yes", "yet",
               "yj", "yl", "you", "youd", "you'd", "you'll", "your", "youre", "you're", "yours", "yourself", "yourselves", "you've", "yr",
               "ys", "yt", "z", "zero", "zi", "zz", "break", "news", "update" , "download", "story", "" ]

#clean text applying all the text preprocessing functions
df0['clean_text'] = df0.text.apply(lambda x: ' '.join(preprocess_text(x)))
df0.head(2)

print(df0.shape)
df0.fillna('', inplace=True)
df0 = df0[df0['clean_text'].ne('')]
df0.shape

df0.reset_index().to_feather('/drive/My Drive/NLP_Project/data/150k_cleantext_df.feather')

### TFIDFing

def top_tfidf_terms(text,tfidf_terms):
  top_terms = []
  tokens = text.split()
  for word in tfidf_terms:
    if word in tokens:
      top_terms.append(word)

    if len(top_terms) == 10:
      break
  
  return top_terms

%%time
from sklearn.feature_extraction.text import TfidfVectorizer

#unigrams

corpus = df0['clean_text']

tfidf=TfidfVectorizer().fit(corpus)
uni_tokens=tfidf.get_feature_names_out()

article_vect = tfidf.transform(corpus)

tf_idf_DF = pd.DataFrame(data=article_vect.tocsr().sum(axis=0),columns=uni_tokens)
tf_idf_DF = tf_idf_DF.T.sort_values(ascending=False, by=[0])  #transposes and sorts the tfidf matrix such that max tfidf terms at top; 

tf_idf_DF['tf-idf'] = tf_idf_DF[0]
tf_idf_DF = tf_idf_DF.drop([0],axis=1)


%%time
df0['tfidf_uni'] = ""
for idx,row in df0.iterrows():
  # print(idx)
  df0.at[idx,'tfidf_uni'] = top_tfidf_terms(row.loc['clean_text'],tf_idf_DF.index)

df0.reset_index().to_feather('/drive/My Drive/NLP_Project/data/150k_uni_df.feather')

def top_tfidf_n_terms(text,n_gram,tfidf_terms):
  # top_terms = {} #if term and value is needed  #if dict then return must be a list of dict [top_term_dict]
  top_terms = []

  tfidf=TfidfVectorizer(ngram_range=(n_gram,n_gram)).fit([text])
  ngram_tokens=tfidf.get_feature_names_out()

  for word in tfidf_terms:
    if word in ngram_tokens:
      top_terms.append(word)
      # top_terms[word] = tf_idf_DF.loc[word][0]

    #only the top n terms are included
    if len(top_terms) == 5:
      break
  
  return top_terms

df0.head(3)

%%time
#bigrams

corpus = df0['clean_text']
tfidf2=TfidfVectorizer(ngram_range=(2,2)).fit(corpus)
n_tokens=tfidf2.get_feature_names_out()
article_vect = tfidf2.transform(corpus)

tf_idf_DF=pd.DataFrame(data=article_vect.tocsr().sum(axis=0),columns=n_tokens)
tf_idf_DF = tf_idf_DF.T.sort_values(ascending=False, by=[0])

tf_idf_DF['tf-idf'] = tf_idf_DF[0]
tf_idf_DF = tf_idf_DF.drop([0],axis=1)


# def top_tfidf_bi_terms(text):
#   # print(text)
#   top_terms = []
#   tfidf=TfidfVectorizer(ngram_range=(2,2)).fit([text])
#   bi_tokens=tfidf.get_feature_names_out()
#   # print(bi_tokens)
#   # tokens = text.split()
#   for word in tf_idf_bi_DF.index:
#     if word in bi_tokens:
#       # tup = (word,tf_idf_DF.loc[word][0])
#       # top_terms.append(tup)
#       top_terms[word] = tf_idf_bi_DF.loc[word][0]

#     if len(top_terms) == 5:
#       # print(top_terms)
#       break
  
#   return [top_terms]

df0.shape

%%time
df0['tfidf_bi'] = ""
for idx,row in df0.iterrows():
  df0.at[idx,'tfidf_bi'] = top_tfidf_n_terms(row.loc['clean_text'],2,tf_idf_DF.index)

df0.reset_index().to_feather('/drive/My Drive/NLP_Project/data/150k_uni_bi_df.feather')

%%time

from sklearn.feature_extraction.text import TfidfVectorizer
#trigrams

corpus = df0['clean_text']
tfidf=TfidfVectorizer(ngram_range=(3,3)).fit(corpus)
n_tokens=tfidf.get_feature_names_out()
article_vect = tfidf.transform(corpus)

tf_idf_DF=pd.DataFrame(data=article_vect.tocsr().sum(axis=0),columns=n_tokens)
tf_idf_DF = tf_idf_DF.T.sort_values(ascending=False, by=[0])

tf_idf_DF['tf-idf'] = tf_idf_DF[0]
tf_idf_DF = tf_idf_DF.drop([0],axis=1)

%%time
df0['tfidf_tri'] = ""
for idx,row in df0.iterrows():
  df0.at[idx,'tfidf_tri'] = top_tfidf_n_terms(row.loc['clean_text'],3,tf_idf_DF.index)

df0.reset_index().to_feather('/drive/My Drive/NLP_Project/data/150k_uni_bi_tri_df.feather')

df0 = pd.read_feather('/drive/My Drive/NLP_Project/data/150k_uni_bi_df.feather')

df0.to_csv('/drive/My Drive/NLP_Project/data/150k_uni_bi_tri_df.csv')
