==> 5_NER.ipynb <==
import time
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

from google.colab import drive
drive.mount('/drive')

df = pd.read_excel("/drive/My Drive/NLP_Project/data/CIL_final_df.xlsx")
df.shape

del df[df.columns[-1]]
del df[df.columns[-1]]
# del df[df.columns[-1]]
del df[df.columns[0]]

df.head(2)

# Spacy NER

# conda install -c conda-forge spacy
!python -m spacy download en_core_web_md

#### RESTART RUNTIME FOR SPACY DOWNLOAD TO WORK

import spacy
spacy_md = spacy.load('en_core_web_md', disable=['parser', 'tagger', 'textcat'])

import nltk
nltk.download('movie_reviews')
nltk.download('punkt')

def ner_spacy(text):
    entities = []
    labels = []

    spacy_doc = spacy_md(text)
    
    for ent in spacy_doc.ents:
        entities.append(ent.text)
        labels.append(ent.label_)

    entities_labels = zip(entities, labels) 
    dataframe = pd.DataFrame({'entities': entities,'labels': labels})
    # dataframe = dataframe[dataframe['labels'].isin(['PERSON','ORG'])]
    dfy = dataframe.groupby('entities').count().sort_values('labels', ascending=False).head(10).reset_index()
    top_ents = dict(dfy.values)

    return top_ents

# text = '''
# great migration pandemic americans flee hightax democratrun states flock republicanled lowtax noinc great migration pandemic americans flee hightax democratrun states flock republicanled lowtax noincometax states study tax foundation people states lowincome tax multitude reasons direct obvious reason reduces tax liability vice president state projects tax foundation jared walczak told yahoo people capacity higher priority analysis contrasted data census bureau released year commercial datasets released week uhaul united van lines low nopersonalincome states utah montana arizona south carolina delaware texas nevada florida north carolina population gains percent district columbia york raise personal income taxes shrunk population pandemic hit americans flocked notax lowincometax states study tax foundation report suggests americans escape burden high income taxes ethan miller works taxes silver spring january york income taxes raised lost percent population july july decline percent population period states individual income taxes states income wages salaries considered taxes vice president state projects tax foundation jared walczak warned ditching hightax states retirees young people flexibility work alaska florida nevada south dakota tennessee texas washington wyoming state income tax district columbia income taxes raised year decline population percent april july census data york lost percent population states highest personal income taxes california hawaii jersey illinois experienced staggering population losses jersey dropped people july july garden states website california rates range percent percent based income population decline period illinois experienced population decline roughly size capital texas states individual income tax gained people percent increase florida grew residents north carolina countrys fourthhighest population growth people south carolina gained residents states population gains implemented enacted cuts individual corporate income tax rates year report walzack warned population country solely blame trend secondorder effects walczak told yahoo states lower tax burdens pro growth tax codes higher rates growth higher economic opportunity – people seek things tax burdens addition reports findings burdening local taxes americans walczak ditching hightax states retirees typically statetostate migration led retirees leaving colder hightax states low income taxes warmer climates walczak emphasized easier young people state lines workfromhome flexibility implemented companies pandemic isnt migration retirees migration people economic opportunities walczak
# '''

# tx = ner_spacy(text)
# type(tx)

# dataframe = pd.DataFrame(tx, columns=['entity', 'label'])
# dataframe = dataframe[dataframe['label'].isin(['PERSON','ORG'])]

# dfy = dataframe.groupby('entity').count().sort_values('label', ascending=False).head(5).reset_index()
# ret = dict(dfy.values)
# ret

df.fillna('', inplace=True)

df['text_ner2'] = df['text'].apply(lambda x: ner_spacy(x))

import re
import sys
from textblob import TextBlob
import nltk
nltk.download('punkt')

df.head(3)

df['cluster_text'] = df.clusters1_uni.map({0:'About Chicago/IL', 4:'CIL;HighTx', 5:'About Chicago/IL', 6:'About Chicago/IL',19:'About Chicago/IL',
                                           26:'About Chicago/IL',18:'Crime',20:'Crime',22:'Crime'})

df.to_excel("/drive/My Drive/NLP_Project/data/CIL_NERandSentiment.xlsx")

df_ner.head(3)

df_ner['cluster_text'].value_counts()

# NERs by Clusters

df = pd.read_excel("/drive/My Drive/NLP_Project/data/CIL_NERandSentiment.xlsx")

df_ner = df[['index1','title','clusters1_uni','text_ner2']]
df_ner.head(3)

# bribe_terms = ['bribe','bribery','corruption','racketeering','michael madigan','dishonest','bribing','fraud']
# df_ner[df_ner['title'].str.lower().str.contains('|'.join(bribe_terms))].shape

# # df_ner = df_ner[df_ner['title'].str.lower().str.contains('|'.join(bribe_terms))]

def str_to_dict(s):
  t = s.replace("'", '"')
  t = t.replace("\\", '')
  res = json.loads(t)
  print(res)
  return res

df_ner['ner_dict2'] = df_ner['text_ner2'].apply(lambda x: str_to_dict(x))

df_ner['cluster_text'].value_counts()

cluster = 20
dfx = df_ner[df_ner['clusters1_uni'] == cluster]

dict_a = {}
for idx,row in dfx.iterrows():
  dict_a.update(row['text_ner2'])

# sorted(dicta.keys())
# sorted(dicta, key=dicta.get)
sorted(dict_a.items(), key=lambda x: x[1], reverse=True)[:30]
