==> 6_Google_CNL_API.ipynb <==
import time
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

from google.colab import drive
drive.mount('/drive')

## Steps 3-7: Analyse the Content Using Cloud Natural Language API

To use the Natural Language API we will import the required libraries.

from google.cloud import language_v1
from google.cloud.language_v1 import enums

import os
os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = '/drive/My Drive/tempp/key.json' 
!echo $GOOGLE_APPLICATION_CREDENTIALS

df = pd.read_excel("/drive/My Drive/tempp/CIL_NERandSentiment.xlsx")
df.shape

df.fillna('', inplace=True)

# dfx = df[['clean_text']].head(5)
df.head(3)

cluster = 0
dfx = df[df['clusters1_uni'] == cluster]

dfx.head()

# df_secondrun = df_secondrun[(df_secondrun['length'] > 1500) & (df_secondrun['length'] < 10000)]


# df_t= df.head(1500)
# del df_t[df_t.columns[0]]
dfx.fillna('', inplace=True)
dfx.shape

# def google_nl_api(text_content):
from datetime import datetime
import time

dfx['g_sent_score']=""
dfx['illinois_entity_sentiment'] = ""
dfx['chicago_entity_sentiment'] = ""
dfx['g_category'] = ""
dfx['g_category_confidence'] = ""


client = language_v1.LanguageServiceClient()
encoding_type = enums.EncodingType.UTF8
type_ = enums.Document.Type.PLAIN_TEXT
language = "en"

start = datetime.now()
time.sleep(1);
counter = 0

for idx,row in dfx.iterrows():

  # while (True):

  #   end = datetime.now()
  #   s = (end-start).seconds
  #   if (counter >= 600):
  #       if (s <= 1):
  #           time.sleep(61) # You can keep track of the time and sleep less, actually
  #           start = datetime.now()
  #           print(f"Just Before Reset:{counter}")
  #           counter = 0

  text_content = row['text'].lower()
  document = {"content": text_content, "type": type_, "language": language}

  # Get sentiment annontations from the document
  annotations = client.analyze_sentiment(document, encoding_type=encoding_type)
  # Get overall document sentiment score
  dfx.loc[idx,'g_sent_score'] = annotations.document_sentiment.score


  #entity_sentiment
  aes_response = client.analyze_entity_sentiment(document, encoding_type=encoding_type)
  # Construct a list of entities where the entity type is a LOCATION
  entities = [entity for entity in aes_response.entities if enums.Entity.Type(entity.type).name == 'LOCATION' and entity.name in ['illinois','chicago']]
  # print(entities)


  
  # Loop through locations
  for entity in entities:
    score_dict =[]
    # print('\n' + '#'*50 + '\n')
    col_name = str(entity.name) + '_entity_sentiment'
    # print(col_name)
    sent = entity.sentiment
    score_dict.append(sent.score)
    score_dict.append(sent.magnitude)
    # score_dict['magnitude'] = sent.magnitude
    dfx.loc[idx,col_name] = str(score_dict)


  # categ ={}
  # category = client.classify_text(document)
  # # print(category.name)
  # # print(category.confidence)
  # # for category in response.categories:
  # #   categ[category.name] = category.confidence
  # #   break

  # # print(categ)
  # dfx.loc[idx,'g_category'] = category.name
  # dfx.loc[idx,'g_category_confidence'] = category.confidence

  end = datetime.now()
  print((end-start).seconds)
  print(f"Normal:{counter}")
  counter += 1

# for idx, row in df_secondrun.iterrows():
#   df_secondrun.loc[idx,'google_sentiment'] = article_sentiments(row['clean_text'])

dfx.to_excel("/drive/My Drive/tempp/google_api_df_0.xlsx")

dfx.head()

## Test cases

articles = df['clean_text'].tolist()
articles

language = "en"
type_ = enums.Document.Type.PLAIN_TEXT


# Analyse the latest 5 articles 
for article in articles[:3]:
  # print (article)
  print('\n' + '#'*50 + '\n')
  # print(article["title"])
  # print(article["link"])
  document = {"content": article, "type": type_, "language": language}
  print_sentiments(document)  
print('\n' + '#'*50)


def sample_classify_text(text_content):
    """
    Classifying Content in a String

    Args:
      text_content The text content to analyze. Must include at least 20 words.
    """

    client = language_v1.LanguageServiceClient()
    encoding_type = enums.EncodingType.UTF8

    # text_content = 'That actor on TV makes movies in Hollywood and also stars in a variety of popular new TV shows.'

    # Available types: PLAIN_TEXT, HTML
    type_ = enums.Document.Type.PLAIN_TEXT

    # Optional. If not specified, the language is automatically detected.
    # For list of supported languages:
    # https://cloud.google.com/natural-language/docs/languages
    language = "en"
    document = {"content": text_content, "type": type_, "language": language}

    response = client.classify_text(document)
    # Loop through classified categories returned from the API
    for category in response.categories:

        print(u"Category name: {}".format(category.name))
        # Get the confidence. Number representing how certain the classifier
        print(u"Confidence: {}".format(category.confidence))


# Analyse the latest 5 articles 
for article in articles:
  print('\n' + '#'*50 + '\n')
  sample_classify_text(article)
 
print('\n' + '#'*50)

articles[:3]
